Add election picker to home view
The behaviour is as such:

If there are other elections in the response,
display the election picker and make another
voterInfo query with the same address when the user selects
another election.

Removes the election picker from view when a voterinfo query
is received that has no other elections in its response.

Added a few tests for VoterInfo/Elections models